+++ lib/life/electricity/service/electricity.dart
@@ -3,9 +3,10 @@ import 'dart:convert';
 import 'package:dio/dio.dart';
 import 'package:mimir/init.dart';
 import 'package:mimir/r.dart';
+import 'package:mimir/session/ywb.dart';
 import '../entity/balance.dart';
 
-const _balanceUrl = "https://xgfy.sit.edu.cn/unifri-flow/WF/Comm/ProcessRequest.do?DoType=DBAccess_RunSQLReturnTable";
+const _balanceUrl = "${YwbSession.base}/unifri-flow/WF/Comm/ProcessRequest.do?DoType=DBAccess_RunSQLReturnTable";
 
 class ElectricityService {
   Dio get _dio => Init.schoolDio;