+++ lib/me/edu_email/service/email.dart
@@ -2,6 +2,8 @@ import 'package:enough_mail/enough_mail.dart';
 import 'package:mimir/credentials/entity/credential.dart';
 import 'package:mimir/credentials/error.dart';
 
+import '../entity/email.dart';
+
 const _server = "imap.mail.sit.edu.cn";
 const _port = 993;
 
@@ -20,9 +22,17 @@ class MailService {
     }
   }
 
-  Future<FetchImapResult> getInboxMessage([int count = 30]) async {
-    final mailboxes = await _client.listMailboxes();
+  Future<List<MailEntity>> getInboxEmails([int count = 50]) async {
+    // final mailboxes = await _client.listMailboxes();
     await _client.selectInbox();
-    return await _client.fetchRecentMessages(messageCount: count);
+    final result = await _client.fetchRecentMessages(messageCount: count);
+    final mails = result.messages.map((msg) => MailEntity.from(msg)).toList();
+    mails.sort((a, b) {
+      final dateA = a.date;
+      final dateB = b.date;
+      if (dateA == null || dateB == null) return 0;
+      return dateA.isAfter(dateB) ? -1 : 1;
+    });
+    return mails;
   }
 }