+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## 2.6.3
 - Migration of new storage structure of timetable and palette.
+- Added UUID in game records.
+- Clear game records on start up.
 
 ## 2.6.2
 - Users are required to accept *Privacy Policy* and *Terms of Services* at the first use.