+++ CHANGELOG.md
@@ -6,7 +6,7 @@
 - Clear game records on start up.
 - Permission requests.
 - Access SIT YWB and electricity bypass VPN.
-- Detecting OA password changed outside of the app.
+- Detecting OA password changed outside the app.
 
 ## 2.6.2
 - Users are required to accept *Privacy Policy* and *Terms of Services* at the first use.